refactor(common)!: move pre-commit validation framework base to org.apache.hudi.core.validator
Relocates BasePreCommitValidator and ValidationContext from
org.apache.hudi.client.validator to org.apache.hudi.core.validator in
hudi-common, continuing the core-package staging for the future
hudi-common/hudi-core module split.

BREAKING CHANGE: both classes shipped in 1.2.0 as
@PublicAPIClass(EVOLVING) and are the extension point for custom
pre-commit validators on the Flink path (hoodie.precommit.validators).
A custom validator compiled against 1.2.0 fails at load with
NoClassDefFoundError and must be recompiled with the new import. This
is an accepted EVOLVING-API change: the config only ever names the
user's concrete class (FQN unchanged), all shipped validators move
together, and no compatibility stub is provided deliberately - a stub
at the old FQN would turn externally compiled overrides of
validateWithMetadata(ValidationContext) into dead overloads once the
parameter type moves, silently skipping the user's validation, which
is worse than the loud load failure.

Engine subclasses and utils in hudi-client-common, hudi-utilities and
hudi-flink update their imports; StreamingOffsetValidator and its test
previously resolved the base types via same-package access and now
import them explicitly. The test also accessed the protected config
field cross-package, which the move surfaced; it now goes through an
accessor on the test's mock subclass.

Verified with the client-common, utilities and flink validator suites.
