+++ hudi-utilities/src/main/java/org/apache/hudi/utilities/streamer/validator/SparkKafkaOffsetValidator.java
@@ -45,7 +45,7 @@ import org.apache.hudi.common.util.CheckpointUtils.CheckpointFormat;
  * <p>This validator is primarily intended for append-only ingestion from Kafka via HoodieStreamer.
  * For upsert workloads with deduplication, configure a higher tolerance or use WARN_LOG.</p>
  *
- * <p><b>Important:</b> This class extends {@link org.apache.hudi.client.validator.BasePreCommitValidator}
+ * <p><b>Important:</b> This class extends {@link org.apache.hudi.core.validator.BasePreCommitValidator}
  * and is invoked by {@link SparkStreamerValidatorUtils}, NOT by {@code SparkValidatorUtils}
  * (which expects {@code SparkPreCommitValidator} with a different constructor signature).
  * Listing this class in {@code hoodie.precommit.validators} while also using the standard
